moveEnums
